// src/app/front-desk/billing/Models/billing.model.ts
export class BillingModel {
    // facilityName: string;
    // providerName: string;
    // visitType: string;
    // visitDate: Date;
    // duration: string;
    // treatmentCode: string;
    // insuranceName: string;
    // claimNo: string;
    // caseType: string;
    // numberOfDays: number;
    // status: string;
    doctorId: string;
    caseId: string;
    patientId: string;
    visitSessionId: string;
    visitType: string;
    visitDate: Date;
    finalize_visit: boolean;
}

export  class BillRecptent{

	type_id: number;
	ids: number;
}